Overrides and Wrappers
Added overrides that allow rules to be applied to change names, params, etc. This was necessary as some functions have params
that are typed with an Enum but really need to be an integer value.

Added in function wrappers methods to wrap the C funs. This doesnt simplify calls to the funs but it does make it possible to
call them without having to specify the `C` library namespace.