Fix for empty structs, crystal docs, com vtables
Empty structs are turned into aliases to `Void`

Wrap the `lib` definition in a conditional to prevent it from being included in the docs because
of the way crystal docs operates. This is a temporary fix until a better solution is found.

Renamed the vtables for com objects to avoid name conflicts with other types.